ClipboardState: moved clipboard-related data/methods to its own state
Todo: also move alerts displayed after copying data to the clipboard
into the clipboard state.
