+++ src/components/shortcuts/KeyboardHotkeys.tsx
@@ -31,7 +31,7 @@ class KeyboardHotkeysClass extends React.Component<WithTranslation> {
     copyTextToClipboard(fileCache: FileState, filesOnly = false): void {
         const length = fileCache.selected.length
 
-        this.appState.copySelectedItemsPath(fileCache, filesOnly)
+        this.appState.clipboard.copySelectedItemsPath(fileCache, filesOnly)
 
         if (length) {
             const { t } = this.injected
